Create networks.sh
the script needs to be run on the the home directory, the df.log will be appended with the information from the commands  thanks to the tee -a command. The df.log can be found in the users home directory. I chose the commands I did because I thought they could be useful. If I find better commands that I would like to use in the future, this file will be simple enough to modify.